+++ tests/roots/test-docstrings/dummy_django_app/models.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from django import VERSION
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
@@ -87,6 +88,14 @@ class ChildModelB(AbstractModel):
     pass
 
 
+def callable_choices():
+    return [("Something", "Not empty")]
+
+
+def callable_choices_empty():
+    return []
+
+
 class ChoiceModel(models.Model):
     choice_limit_below = models.IntegerField(
         choices=[(i, i) for i in range(CHOICES_LIMIT - 1)]
@@ -100,6 +109,9 @@ class ChoiceModel(models.Model):
     choice_with_empty = models.CharField(
         choices=[("", "Empty"), ("Something", "Not empty")]
     )
+    if VERSION >= (5, 0):
+        choice_with_callable = models.CharField(choices=callable_choices)
+        choice_with_callable_empty = models.CharField(choices=callable_choices_empty)
 
 
 class TaggedItem(models.Model):