+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 
 * Add support for Python 3.12
 * Derive the package version from git tags via ``setuptools-scm``
+* [ `#73 <https://github.com/sphinx-doc/sphinxcontrib-django/issues/73>`_ ] Add support for callable choices (`@sevdog <https://github.com/sevdog>`__)
 
 
 Version 2.5 (2023-09-26)